Fix closed guards
**Scope:**

Currently an error causes an issue with guard containers, so with `any_guard` and `all_guards`.

When a guard container contains a guard listening on RUN_EVENT, and another guard listening on custom event, if the first guard closes, the whole guard container will be closed.
This is because currently we make the guard containers closed when all the guards in the current listener map get closed, but we should test if all guards are closed in the whole guard container

**Development:**

Fixed this issue by making it close when all the guards are closed in the whole guard container.